add template&scripts [LogicalDependency_MVI_HANGAR_*_noXVM]
@Willster419
i havent spend a lot of time for the modpack in the last weeks, what also could mean things will not be converted to autoupdatescripts very fast in the near future - consider this only as a first test (also because i couldnt get the local-auto-repo started...).

but right now i´m very glad that automation is here finally!
in the name of everyone who uses and/or maintains this modpack: thank you!!

--
btw from where do you get the uid for new automatinsequences in the root.xml?
i´ve manually typed in some chars... if that´s wrong this needs a fix.

and filenames for scripts MUST have the exact same name, including caps, as the packageName is, right? the app then uses this filename for the {packageName} macro?
nah, at least i did so, cause the other files are named same way...