add automation [Dependency_XVM_XC_vehiclenames]
@Willster419 there might be a problem with the "retrieve_package_property" task.

i wanted to use it in this sequence to get the devUrl from the uid="vxb6fqgbg3p0hfv9" (LogicalDependency_XVM_Main_Nightly) but the return value was the devUrl from Dependency_XVM_Main_Files or LogicalDependency_XVM_Main_Stable

in this sequence there is now a macro for the devurl. this will also work.