Update Dependency_PY_PYmodsCore.xml
this is only a tiny cleanup and a test of the dl-task

@Willster419
i have created an xpath for yadisk and i think i didi it right., but when i use it with this sequence the return value is not the direct-downlaod-url.

from log:
2021-08-06 21:41:06.488   DEBUG: [HtmlWebscrapeParser]: Searching using html path: //a[contains(@class, 'button download-file-button')]//@href
2021-08-06 21:41:06.489   INFO: [HtmlWebscrapeParser]: HtmlPath results in node value '/public/download?hash=OgT1bTA%2FMIx2K0m9V%2Bug5QHn4f0J8zc96ylatWGnnbg%3D&amp;sk=y73fc845626bbcf84b07e41e3837604be' of type 'Attribute'

it must be some kind of uid for a script or something else the app cant handle. as result the download fails.
i´m pretty sure the same happens with gdrive.

also it looks like downloads from forums (kr, wg) are not possible yet. there´s no possibility to enter/save user credentials (and in case this works at some time: we would need a relhax-team-account on forums then...rather than using our own)